Move translation logic and tests to a separate translate module
Extracted 'translate_help_strings' and 'translate_command' along with
their corresponding unit tests from 'main.rs' to a new 'translate'
module ('rust/agama-cli/src/translate.rs') to improve code organization
and maintainability.
